# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: string;
}