# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  serviceId: number;
  serviceName: string;
  servicePrice: number;
  date: string;
  time: string;
  status: BookingStatus;
  notes?: string | null;
  createdAt: string;
}